+++ lit_tests/kernel/wave/index_sequence_analysis.py
@@ -19,6 +19,8 @@ from iree.turbine.kernel.wave.shared_memory_indexing import (
 )
 from iree.turbine.kernel.wave.index_sequence_analysis import (
     partition_strided_operators,
+    set_node_indices,
+    set_post_expansion_indices,
 )
 
 
@@ -84,7 +86,9 @@ def test_gemm():
         IndexingContext.current().finalize()
         promote_placeholders(trace, constraints)
         hoist_allocs(trace)
+        set_node_indices(trace, constraints)
         expand_graph(trace, constraints)
+        set_post_expansion_indices(trace, constraints)
         minimize_global_loads(trace, constraints)
         apply_shared_memory_indexing_corrections(trace, constraints)
         partition_strided_operators(trace, constraints)
